% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhplot-package.r
\docType{package}
\name{mhplot-package}
\alias{mhplot-package}
\alias{mhplot}
\title{Visualise outputs from an optimisation algorithm}
\description{
Visualise outputs from an optimisation algorithm
}
\details{
\tabular{ll}{
Package: \tab mhplot\cr
Type: \tab Package\cr
Version: \tab 0.6 \cr
Date: \tab 2020-08-17 \cr
Release Notes: \tab First release, refactoring out of existing mh package \cr
License: \tab MIT \cr
}

\tabular{lll}{
Version \tab Date \tab Notes \cr
0.5.1 \tab 2015-10-14 \tab First release, refactoring out of existing mh package \cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{calibration}
\keyword{metaheuristics}
\keyword{optimisation}
