% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHplot.R
\name{MHplot}
\alias{MHplot}
\title{Title}
\usage{
MHplot(mat, legend = FALSE)
}
\arguments{
\item{mat}{square contingency table}

\item{legend}{option}
}
\value{
figure
}
\description{
Title
}
\examples{
mat422 <- matrix(c(
0, 10, 10, 10,
30,  0, 10, 10,
30, 30,  0, 10,
30, 30, 30,  0
),4,4,byrow=T)
MHplot(mat422, F)
}
