% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{mkOptimLog}
\alias{mkOptimLog}
\title{Extract information from an optimisation logger}
\usage{
mkOptimLog(
  logDataFrame,
  fitness = "NSE",
  messages = "Message",
  categories = "Category"
)
}
\arguments{
\item{logDataFrame}{a data frame, extracted from an optimisation logger with mh::getLoggerContent}

\item{fitness}{the name of the fitness score that was used for the optimisation, e.g. 'NSE'}

\item{messages}{column name with the content log (messages from optimisation steps)}

\item{categories}{column name with the category of the message (e.g. informative or warnings, etc.)}
}
\value{
an object for graphing through the 'mh' package functions
}
\description{
Extract information from an optimisation logger. This retrieves data and transforms it into a format more amenable to visualisation.
}
